[cocoa] MediaSample with mp3 content can't be deserialised properly. https://bugs.webkit.org/show_bug.cgi?id=300472 rdar://162324480
Reviewed by Eric Carlson.

createFormatDescriptionFromTrackInfo expected all MediaSample but those
with PCM content to have a cookieData; this isn't the case with MP3.
We now only requires the cookieData for Opus and Vorbis which we know for
certain that they requires that information.
For codecs may require it, but that do not have it, it will be handled by
createAudioFormatDescription's call to CMAudioFormatDescriptionCreate

Covered by existing test (which found the problem)
* Source/WebCore/platform/graphics/cocoa/CMUtilities.mm:
(WebCore::audioStreamDescriptionFromAudioInfo):
(WebCore::createFormatDescriptionFromTrackInfo):

Canonical link: https://commits.webkit.org/301295@main