[iOS] Safari crashes underneath `-[UIPreviewTarget initWithContainer:center:transform:]` https://bugs.webkit.org/show_bug.cgi?id=300460 rdar://158751554
Reviewed by Abrar Rahman Protyasha, Tim Horton, and Richard Robinson.

Crashes in Safari are being observed due to a UIKit assertion that checks
whether a targeted preview container is in a window.

It remains unclear exactly how WebKit gets into this state. For now,
speculatively fix by not creating a targeted preview when the container is
not in a window. This should have no user observable effect, since the
container is a subview of `_interactionViewsContainerView`, which is a subview
of the `WKWebView`. Consequently, if the container is not in a window, the
web view must also not be.

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _createTargetedContextMenuHintPreviewIfPossible]):

Canonical link: https://commits.webkit.org/301302@main