Many PDF tests time out sporadically as a result of the PDF not loading https://bugs.webkit.org/show_bug.cgi?id=296096 rdar://156008706
Reviewed by Abrar Rahman Protyasha.

ResourceLoader has a fast path, `deliverResponseAndData`, which seems - at least
for PDFs - to be somewhat rarely encountered in practice. However, with small
test PDFs loaded from the filesystem, it is hit enough to cause reproducible flaky test failures.

This fast path calls `didReceiveBuffer` to hand the data to the client, and
then immediately `didFinishLoading`.

Currently NetscapePlugInStreamLoader overrides ResourceLoader's `didReceiveData`,
but not `didReceiveBuffer`, in order to slurp the data aside for the plugin.
This bug stems from this fact, and the fact that `deliverResponseAndData`
does not call `didReceiveData`, skipping the override.

Since `didReceiveBuffer` now appears to be the bottleneck for all paths,
replace NetscapePlugInStreamLoader's `didReceiveData` override with an override
of this method. Thus, the PDF plugin now intercepts the data when we arrive via
`deliverResponseAndData`, where it wouldn't before.

No new tests, covered by existing tests that were previously expected to be flaky.

* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/loader/NetscapePlugInStreamLoader.cpp:
(WebCore::NetscapePlugInStreamLoader::didReceiveBuffer):
(WebCore::NetscapePlugInStreamLoader::didReceiveData): Deleted.
* Source/WebCore/loader/NetscapePlugInStreamLoader.h:

Canonical link: https://commits.webkit.org/301346@main