[JSC] Stop calling Promise.resolve from C++ code https://bugs.webkit.org/show_bug.cgi?id=300572 rdar://162454169
Reviewed by Sosuke Suzuki.

This patch implements PromiseResolve abstract operation in C++ and
JSPromise::resolvedPromise calls it instead of calling JS function
Promise.resolve. But we are not still changing Promise.resolve from JS
to C++ because that can be a regression when Promise.resolve is inlined
and most of code is removed via that. We should have DFG intrinsic for
this Promise.resolve and we need to lower it into optimal form in DFG /
FTL. So until that, we cannot move it to C++.

* Source/JavaScriptCore/runtime/JSGlobalObject.cpp:
(JSC::JSGlobalObject::init):
(JSC::JSGlobalObject::visitChildrenImpl):
* Source/JavaScriptCore/runtime/JSGlobalObject.h:
(JSC::JSGlobalObject::promiseResolveFunction const): Deleted.
* Source/JavaScriptCore/runtime/JSPromise.cpp:
(JSC::JSPromise::resolvedPromise):
(JSC::JSPromise::promiseResolve):
* Source/JavaScriptCore/runtime/JSPromise.h:

Canonical link: https://commits.webkit.org/301379@main