[CoordinatedGraphics] Compute the layers tile size instead of using a fixed value https://bugs.webkit.org/show_bug.cgi?id=300093
Reviewed by Miguel Gomez.

This way we can use a different logic depending on whether we are using
CPU or GPU rendering for the tiles. In the case of CPU rendering we use
smaller tiles of 256x256 by default, while for GPU rendering we use the
visible size for the width and 1/4 of the visible height, similar to what
chromium does.

Canonical link: https://commits.webkit.org/301398@main