REGRESSION(298715@main): SVG animation with finite 'max' and indefinite 'repeatDur` causes the web page to jetsam https://bugs.webkit.org/show_bug.cgi?id=299047 rdar://160099991
Reviewed by Nikolas Zimmermann.

Re-landing 300165@main after confirming it did not cause a memory regression.

When the SMIL ends because the elapsed time is greater than the intervalEnd, we
should not use repeatingDuration to calculate 'repeat' value unless its value is
finite. If it is value isIndefinite, we should use the duration since the begin
of the animation.

Test: svg/animations/smil-max-finite-repeatDur-indefinite-crash.html

* LayoutTests/svg/animations/smil-max-finite-repeatDur-indefinite-crash-expected.txt: Added.
* LayoutTests/svg/animations/smil-max-finite-repeatDur-indefinite-crash.html: Added.
* Source/WebCore/svg/animation/SVGSMILElement.cpp:
(WebCore::SVGSMILElement::calculateAnimationPercentAndRepeat const):

Canonical link: https://commits.webkit.org/301404@main