[JSC] WASM IPInt SIMD: explicitly skip JIT-less test modes for SIMD tests that require JIT https://bugs.webkit.org/show_bug.cgi?id=300517 rdar://162381319
Reviewed by Yusuke Suzuki.

Soon we will set useWasmIPIntSIMD to true by default. When doing so,
I will remove the code in run-jsc-stress-tests that skips the
wasm-no-jit and wasm-no-wasm-jit modes for tests that use SIMD
since those tests should now be runnable using IPInt.

However, there are a few tests that we'll continue to skip when
JIT is disabled. Mark those explicitly now to prepare.

While inline-simd-function.js, ipint-bbq-osr-with-try3.js,
and ipint-bbq-osr-with-try4.js aren't explicitly marked with
@requireOptions("--useWasmSIMD=1"), the test code itself catches
the CompileError that is thrown when running with JIT disabled and
IPInt SIMD disabled. So skipping these JIT-less modes is not a loss
of test coverage but will be necessary once IPInt SIMD is enabled
because these tests run too slowly without JIT.

armv7-simple-loop-osr.js explicitly checks that it tiered up to OMG,
so lets explicitly skip that when JIT is disabled.

* JSTests/wasm/stress/armv7-simple-loop-osr.js:
* JSTests/wasm/stress/inline-simd-function.js:
* JSTests/wasm/stress/ipint-bbq-osr-with-try3.js:
* JSTests/wasm/stress/ipint-bbq-osr-with-try4.js:

Canonical link: https://commits.webkit.org/301410@main