Add a test for 301346@main, ensuring that loading cached PDFs succeeds https://bugs.webkit.org/show_bug.cgi?id=300662 rdar://162560992
Reviewed by Sam Weinig and Abrar Rahman Protyasha.

Upon further investigation, it is trivial to make a 100% repro test
case for the bug fixed in 301346@main, where PDFPlugin would fail to
load PDFs that were already cached, so that we don't have to depend
on a rare race condition to cover that fix.

It turns out that simply waiting for PDFPlugin to itself force-cache
the loaded PDF via PDFPluginBase::addArchiveResource(), and then trying
to load the *same* PDF in another plugin, will take the archive path
(and thus the `deliverResponseAndData` path, and thus exercise the fix
from 301346@main), 100% of the time.

* LayoutTests/pdf/load-pdf-twice-in-sequence-expected.txt: Added.
* LayoutTests/pdf/load-pdf-twice-in-sequence.html: Added.

Canonical link: https://commits.webkit.org/301453@main