[libpas] Switch to open source MTE implementation for all platforms https://bugs.webkit.org/show_bug.cgi?id=300555 rdar://162125588
Reviewed by Yusuke Suzuki.

Though the actual MTE enablement code is checked at runtime (and so is
not dependent on platform -- we can build it even for non-MTE devices),
for the purposes of smoothing this transition we wanted to first enable
it for iOS devices (which comprise a superset of those Apple SoCs which
support the feature) in case there are any platform-specific build or
performance issues with the alternative open source MTE implementation.

Now that the patch enabling this build for iOS devices (301361@main) has
gone in without issues, we can move on to the next step of enabling this
build generally. This does not necessarily enable MTE itself on all
devices, as there remains the runtime check that determines whether or
not the allocator actually uses MTE, but it does mean that all MTE code
(including that runtime check) would be built from the WebKit sources.

Canonical link: https://commits.webkit.org/301490@main