[build.webkit.org] Build requests are not collapsed anymore since the upgrade of buildbot to 4.3.0 https://bugs.webkit.org/show_bug.cgi?id=300713
Reviewed by Aakash Jain.

It seems the default value of this parameter has changed between buildbot versions,
and we want it enabled for build.webkit.org.

This allows buildbot to always pick the last build-request (the last commit on main)
when there are several waiting to be built, collapsing the others into this last build
request.

* Tools/CISupport/build-webkit-org-webserver/master.cfg:
* Tools/CISupport/build-webkit-org/master.cfg:

Canonical link: https://commits.webkit.org/301496@main