[GTK] Move ValidationBubble implementation from WebCore to WebKit https://bugs.webkit.org/show_bug.cgi?id=300618
Reviewed by Michael Catanzaro.

It's only used by WebKit in the UI process, so we don't need it in
WebCore. We can't move the base class to WebKit yet because it's also
used by WebKitLegacy, so for now this patch makes the base class
asbtract for GTK port and uses a derived class from WebKit.

Canonical link: https://commits.webkit.org/301538@main