[libpas] Build fails with libc implementations that lack execinfo.h https://bugs.webkit.org/show_bug.cgi?id=300701
Reviewed by Yusuke Suzuki and Michael Catanzaro.

Change guards to use backtrace() and execinfo.h on Linux only when using
glibc as the C library. The PlayStation and Windows cases no longer need
to be matched, as they are neither Linux nor Darwin, both of which are now
explicitly listed. The Android check is kept as it was.

* Source/bmalloc/libpas/src/libpas/pas_probabilistic_guard_malloc_allocator.c:
* Source/bmalloc/libpas/src/test/PGMTests.cpp:
(addPGMTests):

Canonical link: https://commits.webkit.org/301539@main