[JSC] Wasm: fix register usage when JIT-less throws TypeErrorInvalidValueUse https://bugs.webkit.org/show_bug.cgi?id=300751 rdar://162643391
Reviewed by Yusuke Suzuki.

Arguments should be in a0, a1, not r0, r1. They happen to be the
same on ARM64 but that's not the case on X86_64, so this bug is apparent
only on X86_64.

Also note that this part of the JIT-less path is probably dead today
since today we require JIT to run Wasm SIMD, which means any mention
of v128 in JIT-less is a compile time error. However, once IPInt SIMD
support is enabled, this part of the path will become used.

* Source/JavaScriptCore/llint/InPlaceInterpreter.asm:

Canonical link: https://commits.webkit.org/301559@main