[MiniBrowser] Fix parameter passing to WebKitTestRunner. https://bugs.webkit.org/show_bug.cgi?id=300809 rdar://162689631
Reviewed by Sihui Liu.

We can pass debugging flag like this before:

$ run-minibrowser -WebCoreLogging Layout

But it regressed at 299873@main. This is because we need to pass as is to WKTR.

At 299873@main, intension was to allow this kind of order:

$ run-minibrowser --site-isolation URL
$ run-minibrowser URL --site-isolation

To make this happen, we need to add `--url` before URL to WKTR. For known arguments,
this is perfect. But this doesn't work for the unknown arguments for python script.

So to support this kind of arguments, we cannot write like this:

$ run-minibrowser -WebCoreLogging Layout URL

URL must come before debug flags.

* Tools/Scripts/webkitpy/minibrowser/run_webkit_app.py:
(main):

Canonical link: https://commits.webkit.org/301582@main