Update LLVM (2025-10-15) https://bugs.webkit.org/show_bug.cgi?id=300841
Reviewed by Geoffrey Garen and Chris Dumez.

Rollout the latest WebKit branch of LLVM for static analysis.

This rollout includes the fix to treat the return value of NSStringFromSelector and related
functions as safe, and fix for ForwardDeclChecker to ignore unary operator `&` and `*` when
considering if a given function argument is also a function argument in the caller / call-site.

* Tools/CISupport/safer-cpp-llvm-version:

Canonical link: https://commits.webkit.org/301588@main