Safari Technology Preview 226 breaks espn.com title bar https://bugs.webkit.org/show_bug.cgi?id=297721 rdar://158951219
Reviewed by Chris Dumez and Anne van Kesteren.

With the Navigation API enabled, the ESPN title bar is empty instead of
containing the tab buttons like it should.

The issue is that espn.com re-assigns the "navigation" variable to hold
these items in this function:

preRender: function(type, navCached, defaultNavData) {
    ...
    navigation = defaultNavData.navigation
    ...
    espn_ui.Helpers.nav.items = navigation.items
    ...
}

But with the Navigation API enabled, the "navigation" variable is reserved
for accessing that API and is not replaceable. So this re-assignment fails
and the title bar ends up empty.

The spec does not say that "navigation" should be replaceable. But this is
breaking a live site and Chrome does allow it to be replaceable, so we fix this
by making it replaceable in WebKit too. I have confirmed that this does fix
the ESPN title bar locally.

Associated spec issue: https://github.com/whatwg/html/pull/11786

I've added a new layout test to check this:
navigation-api/navigation-object-is-replaceable.html

* LayoutTests/imported/w3c/web-platform-tests/interfaces/html.idl:
* LayoutTests/imported/w3c/web-platform-tests/interfaces/html.idl:
* LayoutTests/navigation-api/navigation-object-is-replaceable-expected.txt: Added.
* LayoutTests/navigation-api/navigation-object-is-replaceable.html: Added.
* Source/WebCore/page/DOMWindow.idl:

Canonical link: https://commits.webkit.org/301607@main