AX: AccessibilityNodeObject::clearChildren() can dereference a nullptr https://bugs.webkit.org/show_bug.cgi?id=300851 rdar://161093750
Reviewed by Chris Fleizach.

Given this sequence:

  1. Element::~Element()
  2. AXObjectCache::remove(WebCore::Node&)
  3. AXObjectCache::remove(AXID)
  4. AccessibilityNodeObject::detachRemoteParts()
  5. AccessibilityNodeObject::clearChildren()
  6. AccessibilityNodeObject::isNativeLabel()

We dereference a nullptr when trying to check dynamicDowncast<HTMLLabelElement>(node()), because while the node is non-null
and not yet destroyed, it's m_tagName field has been cleared (and is this null), so the is<HTMLLabelElement> check dereferences
the null tagname.

Avoid this by removing the isNativeLabel() branch entirely. This virtual function call is not necessary, we can just
unconditionally reset the m_containsOnlyStaticText and m_containsOnlyStaticTextDirty bools. Also remove unnecessary
virtual function call to isTable() to avoid similar issues.

* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::clearChildren):

Canonical link: https://commits.webkit.org/301612@main