[build.webkit.org] Build requests are not collapsed since the switch to multi-master mode https://bugs.webkit.org/show_bug.cgi?id=300920
Reviewed by Aakash Jain.

At 301496@main I thought that the issue with the build requests not being collapsed anymore
since the migration to version 4.3 was because a change of default value for the parameter
collapseRequests from 2.10 to 4.3, but this was not the case in the end.

The problem is not caused by the new version, but by the new multi-master deployment.
The "-webserver" master is configured without the list of builders and workers, and that
causes that the feature to collapse the requests (which runs on this master) to be unable
to find the object representing the builder (given its name), and then it is skipping to
run the function to collapse the builds which is part of this builder object.

Fix that by defining also the list of workers for the "-webserver" master, that way
it can find the builder object and the feature to collapse build requests works again.

This also reverts 301496@main since the value to collapseRequests actually defauls to True.

* Tools/CISupport/build-webkit-org-webserver/master.cfg:
* Tools/CISupport/build-webkit-org/master.cfg:

Canonical link: https://commits.webkit.org/301655@main