Reduce unsafeness in WebCore/editing https://bugs.webkit.org/show_bug.cgi?id=300869
Reviewed by Ryosuke Niwa.

Apply https://github.com/WebKit/WebKit/wiki/Safer-CPP-Guidelines to a
variety of files.

SelectionGeometryGatherer.cpp includes a subtle change. We go from

  m_gatherer.m_renderView->view().frame()

to

  m_gatherer.m_renderView->frame()

in order to be able to reuse a variable. As far as I can tell these are
equivalent so this should not be observable.

Canonical link: https://commits.webkit.org/301692@main