[GTK][WPE] Fix tracking and leaking reallocated memory with MALLOC_HEAP_BREAKDOWN https://bugs.webkit.org/show_bug.cgi?id=300882
Reviewed by Michael Catanzaro.

Memory reallocated with zoneRealloc() was not tracked properly.

If realloc() returns a different pointer, the old pointer is left
accounted for in the zone allocations with the new size, and the new
pointer is not tracked at all.

So when the new pointer is freed by zoneFree():
- the new memory was leaked because the function checks if the
  allocation is tracked before calling free()
- the old pointer allocation was left in the allocation statistics

This has been noticed with a backport of this feature on WPE 2.42,
which showed incorrect leaks of AssemblerData and MetadataTable objects.

* Source/WTF/wtf/malloc_heap_breakdown/main.cpp:
(MallocZoneHeapManager::zoneRealloc):

Canonical link: https://commits.webkit.org/301709@main